Colocate some methods
