+++ app/containers/EmojiPicker/index.tsx
@@ -6,12 +6,12 @@ import TabBar from './TabBar';
 import EmojiCategory from './EmojiCategory';
 import Footer from './Footer';
 import styles from './styles';
-import { categories, emojisByCategory } from './data';
+import { categories, emojisByCategory } from '../../lib/constants';
 import { useTheme } from '../../theme';
 import { IEmoji, ICustomEmojis } from '../../definitions';
-import { useAppSelector } from '../../lib/hooks';
+import { useAppSelector, useFrequentlyUsedEmoji } from '../../lib/hooks';
+import { addFrequentlyUsed } from '../../lib/methods';
 import { IEmojiPickerProps, EventTypes } from './interfaces';
-import { useFrequentlyUsedEmoji, addFrequentlyUsed } from './frequentlyUsedEmojis';
 
 const EmojiPicker = ({
 	onItemClicked,
