+++ app/lib/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useAppSelector';
 export * from './usePermissions';
+export * from './useFrequentlyUsedEmoji';
