+++ app/views/RoomView/ReactionPicker.tsx
@@ -5,7 +5,7 @@ import EmojiPicker from '../../containers/EmojiPicker';
 import styles from './styles';
 import { IEmoji } from '../../definitions';
 import { EventTypes } from '../../containers/EmojiPicker/interfaces';
-import { searchEmojis } from '../../containers/EmojiPicker/helpers';
+import { searchEmojis } from '../../lib/methods';
 import { useDebounce } from '../../lib/methods/helpers/debounce';
 import { EmojiSearch } from '../../containers/EmojiPicker/EmojiSearch';
 import { events, logEvent } from '../../lib/methods/helpers/log';
